+++ Posterizarr.ps1
@@ -8,7 +8,7 @@ param (
     [string]$mediatype
 )
 
-$CurrentScriptVersion = "1.2.44"
+$CurrentScriptVersion = "1.2.47"
 $global:HeaderWritten = $false
 $ProgressPreference = 'SilentlyContinue'
 
@@ -658,7 +658,12 @@ function GetTMDBMoviePoster {
                 $NoLangPoster = ($response.images.posters | Where-Object iso_639_1 -eq $null)
                 if (!$NoLangPoster) {
                     if (!$global:OnlyTextless -eq 'true') {
-                        $posterpath = (($response.images.posters | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($response.images.posters | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($response.images.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($global:FavProvider -eq 'TMDB') {
                             $global:Fallback = "fanart"
@@ -666,7 +671,12 @@ function GetTMDBMoviePoster {
                         }
                         Write-Entry -Subtext "Found Poster with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                         $global:PosterWithText = $true
-                        $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object vote_average -Descending)[0]).iso_639_1
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object -Descending)[0]).iso_639_1
+                        }
+                        Else {
+                            $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                        }
                         $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/movie/$($global:tmdbid)/images/posters"
                     }
                     Else {
@@ -675,7 +685,12 @@ function GetTMDBMoviePoster {
                     }
                 }
                 Else {
-                    $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                    if ($global:TMDBVoteSorting -eq 'Primary'){
+                        $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                    }
+                    Else {
+                        $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                    }
                     if ($posterpath) {
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Textless Poster on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -712,7 +727,12 @@ function GetTMDBMoviePoster {
                         $FavPoster = ($response.images.posters | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found Poster without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -757,7 +777,12 @@ function GetTMDBMovieBackground {
                 $NoLangPoster = ($response.images.backdrops | Where-Object iso_639_1 -eq $null)
                 if (!$NoLangPoster) {
                     if (!$global:OnlyTextless -eq 'true') {
-                        $posterpath = (($response.images.backdrops | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($response.images.backdrops | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($response.images.backdrops | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($global:FavProvider -eq 'TMDB') {
                             $global:Fallback = "fanart"
@@ -765,7 +790,12 @@ function GetTMDBMovieBackground {
                         }
                         Write-Entry -Subtext "Found background with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                         $global:PosterWithText = $true
-                        $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object vote_average -Descending)[0]).iso_639_1
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object -Descending)[0]).iso_639_1
+                        }
+                        Else {
+                            $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                        }
                         $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/movie/$($global:tmdbid)/images/backdrops"
                     }
                     Else {
@@ -774,7 +804,12 @@ function GetTMDBMovieBackground {
                     }
                 }
                 Else {
-                    $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                    if ($global:TMDBVoteSorting -eq 'Primary'){
+                        $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                    }
+                    Else {
+                        $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                    }
                     if ($posterpath) {
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Textless background on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -818,7 +853,12 @@ function GetTMDBMovieBackground {
                         $FavPoster = ($response.images.backdrops | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found background without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -874,7 +914,12 @@ function GetTMDBShowPoster {
                 $NoLangPoster = ($response.images.posters | Where-Object iso_639_1 -eq $null)
                 if (!$NoLangPoster) {
                     if (!$global:OnlyTextless -eq 'true') {
-                        $posterpath = (($response.images.posters | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($response.images.posters | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($response.images.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($global:FavProvider -eq 'TMDB') {
                             $global:Fallback = "fanart"
@@ -882,7 +927,12 @@ function GetTMDBShowPoster {
                         }
                         Write-Entry -Subtext "Found Poster with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                         $global:PosterWithText = $true
-                        $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object vote_average -Descending)[0]).iso_639_1
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object -Descending)[0]).iso_639_1
+                        }
+                        Else {
+                            $global:TMDBAssetTextLang = (($response.images.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                        }
                         $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/tv/$($global:tmdbid)/images/posters"
                         return $global:posterurl
                     }
@@ -892,7 +942,12 @@ function GetTMDBShowPoster {
                     }
                 }
                 Else {
-                    $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                    if ($global:TMDBVoteSorting -eq 'Primary'){
+                        $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                    }
+                    Else {
+                        $posterpath = (($response.images.posters | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                    }
                     if ($posterpath) {
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Textless Poster on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -929,7 +984,12 @@ function GetTMDBShowPoster {
                         $FavPoster = ($response.images.posters | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found Poster without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -975,11 +1035,21 @@ function GetTMDBSeasonPoster {
                 Write-Entry -Subtext "NoLangPoster: $NoLangPoster" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Cyan -log Debug
                 if (!$NoLangPoster) {
                     if (!$global:SeasonOnlyTextless) {
-                        $posterpath = (($response.posters | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($response.posters | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($response.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Season Poster with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                         $global:PosterWithText = $true
-                        $global:TMDBAssetTextLang = (($response.posters | Sort-Object vote_average -Descending)[0]).iso_639_1
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $global:TMDBAssetTextLang = (($response.posters | Sort-Object -Descending)[0]).iso_639_1
+                        }
+                        Else {
+                            $global:TMDBAssetTextLang = (($response.posters | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                        }
                         $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/tv/$($global:tmdbid)/season/$global:SeasonNumber/images/posters"
                         $global:TMDBSeasonFallback = $global:posterurl
                         Write-Entry -Subtext "Posterpath: $posterpath" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Cyan -log Debug
@@ -996,7 +1066,12 @@ function GetTMDBSeasonPoster {
                     }
                 }
                 Else {
-                    $posterpath = (($response.posters | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                    if ($global:TMDBVoteSorting -eq 'Primary'){
+                        $posterpath = (($response.posters | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                    }
+                    Else {
+                        $posterpath = (($response.posters | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                    }
                     if ($posterpath) {
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Textless Season Poster on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -1047,7 +1122,12 @@ function GetTMDBSeasonPoster {
                         $FavPoster = ($responseBackup.images.posters | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found Poster without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -1081,7 +1161,12 @@ function GetTMDBSeasonPoster {
                         $FavPoster = ($response.posters | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found Poster without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -1131,7 +1216,12 @@ function GetTMDBShowBackground {
                 $NoLangPoster = ($response.images.backdrops | Where-Object iso_639_1 -eq $null)
                 if (!$NoLangPoster) {
                     if (!$global:OnlyTextless -eq 'true') {
-                        $posterpath = (($response.images.backdrops | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($response.images.backdrops | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($response.images.backdrops | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($global:FavProvider -eq 'TMDB') {
                             $global:Fallback = "fanart"
@@ -1139,7 +1229,12 @@ function GetTMDBShowBackground {
                         }
                         Write-Entry -Subtext "Found background with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                         $global:PosterWithText = $true
-                        $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object vote_average -Descending)[0]).iso_639_1
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object -Descending)[0]).iso_639_1
+                        }
+                        Else {
+                            $global:TMDBAssetTextLang = (($response.images.backdrops | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                        }
                         $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/tv/$($global:tmdbid)/images/backdrops"
                     }
                     Else {
@@ -1148,7 +1243,12 @@ function GetTMDBShowBackground {
                     }
                 }
                 Else {
-                    $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                    if ($global:TMDBVoteSorting -eq 'Primary'){
+                        $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                    }
+                    Else {
+                        $posterpath = (($response.images.backdrops | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                    }
                     if ($posterpath) {
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         Write-Entry -Subtext "Found Textless background on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -1196,7 +1296,12 @@ function GetTMDBShowBackground {
                         $FavPoster = ($response.images.backdrops | Where-Object iso_639_1 -eq $lang)
                     }
                     if ($FavPoster) {
-                        $posterpath = (($FavPoster | Sort-Object vote_average -Descending)[0]).file_path
+                        if ($global:TMDBVoteSorting -eq 'Primary'){
+                            $posterpath = (($FavPoster | Sort-Object -Descending)[0]).file_path
+                        }
+                        Else {
+                            $posterpath = (($FavPoster | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                        }
                         $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                         if ($lang -eq 'null') {
                             Write-Entry -Subtext "Found background without Language on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
@@ -1251,16 +1356,31 @@ function GetTMDBTitleCard {
         if ($response.stills) {
             $NoLangPoster = ($response.stills | Where-Object iso_639_1 -eq $null)
             if (!$NoLangPoster) {
-                $posterpath = (($response.stills | Sort-Object vote_average -Descending)[0]).file_path
+                if ($global:TMDBVoteSorting -eq 'Primary'){
+                    $posterpath = (($response.stills | Sort-Object -Descending)[0]).file_path
+                }
+                Else {
+                    $posterpath = (($response.stills | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                }
                 $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                 Write-Entry -Subtext "Found Title Card with text on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
                 $global:PosterWithText = $true
-                $global:TMDBAssetTextLang = (($response.stills | Sort-Object vote_average -Descending)[0]).iso_639_1
+                if ($global:TMDBVoteSorting -eq 'Primary'){
+                    $global:TMDBAssetTextLang = (($response.stills | Sort-Object -Descending)[0]).iso_639_1
+                }
+                Else {
+                    $global:TMDBAssetTextLang = (($response.stills | Sort-Object $global:TMDBVoteSorting -Descending)[0]).iso_639_1
+                }
                 $global:TMDBAssetChangeUrl = "https://www.themoviedb.org/tv/$($global:tmdbid)/season/$global:season_number/episode/$global:episodenumber/images/backdrops"
                 return $global:posterurl
             }
             Else {
-                $posterpath = (($response.stills | Where-Object iso_639_1 -eq $null | Sort-Object vote_average -Descending)[0]).file_path
+                if ($global:TMDBVoteSorting -eq 'Primary'){
+                    $posterpath = (($response.stills | Where-Object iso_639_1 -eq $null | Sort-Object -Descending)[0]).file_path
+                }
+                Else {
+                    $posterpath = (($response.stills | Where-Object iso_639_1 -eq $null | Sort-Object $global:TMDBVoteSorting -Descending)[0]).file_path
+                }
                 if ($posterpath) {
                     $global:posterurl = "https://image.tmdb.org/t/p/original$posterpath"
                     Write-Entry -Subtext "Found Textless Title Card on TMDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Green -log Info
@@ -1841,10 +1961,44 @@ function GetTVDBMovieBackground {
             }
             if ($response) {
                 if ($response.data.artworks) {
-                    $global:posterurl = ($response.data.artworks | Where-Object { $_.language -eq $null -and $_.type -eq '15' } | Sort-Object Score)[0].image
-                    Write-Entry -Subtext "Found Textless Background on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
-                    $global:TVDBAssetChangeUrl = "https://thetvdb.com/movies/$($response.data.slug)#artwork"
-                    return $global:posterurl
+                    $NoLangArtwork = $response.data.artworks | Where-Object { $null -eq $_.language -and $_.type -eq '15' }
+                    if ($NoLangArtwork){
+                        $global:posterurl = ($NoLangArtwork | Sort-Object Score)[0].image
+                        Write-Entry -Subtext "Found Textless Background on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
+                        $global:TVDBAssetChangeUrl = "https://thetvdb.com/movies/$($response.data.slug)#artwork"
+                        return $global:posterurl
+                    }
+                    Else {
+                        # Trying other languages
+                        foreach ($lang in $global:PreferredLanguageOrderTVDB) {
+                            if ($lang -eq 'null') {
+                                $LangArtwork = ($response.data.artworks | Where-Object { $_.language -like "" -and $_.type -eq '15' } | Sort-Object Score)
+                            }
+                            Else {
+                                $LangArtwork = ($response.data.artworks | Where-Object { $_.language -like "$lang*" -and $_.type -eq '15' } | Sort-Object Score)
+                            }
+                            if ($LangArtwork) {
+                                $global:posterurl = $LangArtwork[0].image
+                                if ($lang -eq 'null') {
+                                    Write-Entry -Subtext "Found Background without Language on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
+                                }
+                                Else {
+                                    Write-Entry -Subtext "Found Background with Language '$lang' on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Blue -log Info
+                                }
+                                if ($lang -ne 'null') {
+                                    $global:PosterWithText = $true
+                                    $global:TVDBAssetTextLang = $lang
+                                }
+                                $global:TVDBAssetChangeUrl = "https://thetvdb.com/movies/$($response.data.slug)#artwork"
+                                return $global:posterurl
+                                continue
+                            }
+                        }
+                        if (!$global:posterurl) {
+                            Write-Entry -Subtext "No background found on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Yellow -log Warning
+                            $global:TVDBAssetChangeUrl = "https://thetvdb.com/movies/$($response.data.slug)#artwork"
+                        }
+                    }
                 }
                 Else {
                     Write-Entry -Subtext "No Background found on TVDB" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Yellow -log Warning
@@ -2581,6 +2735,7 @@ function LogConfigSettings {
         Write-Entry -Subtext "| Plex Token:                   $($PlexToken[0..7] -join '')****" -Path $configLogging -Color White -log Info
     }
     Write-Entry -Subtext "| Fav Provider:                 $global:FavProvider" -Path $configLogging -Color White -log Info
+    Write-Entry -Subtext "| TMDB Vote Sorting:            $global:TMDBVoteSorting" -Path $configLogging -Color White -log Info
     Write-Entry -Subtext "| Preferred Lang Order:         $($global:PreferredLanguageOrder -join ',')" -Path $configLogging -Color White -log Info
     Write-Entry -Subtext "| Preferred Season Lang Order:  $($global:PreferredSeasonLanguageOrder -join ',')" -Path $configLogging -Color White -log Info
     Write-Entry -Subtext "Plex Part" -Path $configLogging -Color Cyan -log Info
@@ -3037,6 +3192,11 @@ $global:tmdbtoken = $config.ApiPart.tmdbtoken
 $FanartTvAPIKey = $config.ApiPart.FanartTvAPIKey
 $PlexToken = $config.ApiPart.PlexToken
 $global:FavProvider = $config.ApiPart.FavProvider.ToUpper()
+$global:TMDBVoteSorting = $config.ApiPart.tmdb_vote_sorting
+if (!$global:TMDBVoteSorting) {
+    Write-Entry -Message "TMDB Sorting option not set in config, setting it to 'vote_average' for you" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Yellow -log Warning
+    $global:TMDBVoteSorting = "vote_average"
+}
 $global:PreferredLanguageOrder = $config.ApiPart.PreferredLanguageOrder
 $global:PreferredSeasonLanguageOrder = $config.ApiPart.PreferredSeasonLanguageOrder
 
@@ -11027,27 +11187,29 @@ else {
 
         # Perform deletion of unchecked items
         foreach ($uncheckedItem in $uncheckedItems) {
-            # Full path to the item
-            $uncheckedItemPath = $uncheckedItem + ".jpg"
+            if ($uncheckedItem -notlike '*.jpg') {
+                # Full path to the item
+                $uncheckedItemPath = $uncheckedItem + ".jpg"
 
-            # Check if its a asset from Posterizarr
-            $exifmagickcommand = "& `"$magick`" identify -verbose `"$uncheckedItemPath`""
-            $exifmagickcommand | Out-File $global:ScriptRoot\Logs\ImageMagickCommands.log -Append
+                # Check if its a asset from Posterizarr
+                $exifmagickcommand = "& `"$magick`" identify -verbose `"$uncheckedItemPath`""
+                $exifmagickcommand | Out-File $global:ScriptRoot\Logs\ImageMagickCommands.log -Append
 
-            $ExifData = (Invoke-Expression $exifmagickcommand | Select-String -Pattern 'created with ppm|created with posterizarr')
+                $ExifData = (Invoke-Expression $exifmagickcommand | Select-String -Pattern 'created with ppm|created with posterizarr')
 
-            if ($ExifData) {
-                # Remove unchecked item from filesystem
-                Remove-Item -LiteralPath $uncheckedItemPath -Force | Out-Null
-                $ImagesCleared++
-                Write-Entry -Subtext "Artwork Removed: $uncheckedItemPath" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Yellow -log Info
-                if ($LibraryFolders -eq 'true') {
-                    # Determine the parent directory of the item
-                    $parentDir = Split-Path -Path $uncheckedItemPath -Parent
+                if ($ExifData) {
+                    # Remove unchecked item from filesystem
+                    Remove-Item -LiteralPath $uncheckedItemPath -Force | Out-Null
+                    $ImagesCleared++
+                    Write-Entry -Subtext "Artwork Removed: $uncheckedItemPath" -Path $global:ScriptRoot\Logs\Scriptlog.log -Color Yellow -log Info
+                    if ($LibraryFolders -eq 'true') {
+                        # Determine the parent directory of the item
+                        $parentDir = Split-Path -Path $uncheckedItemPath -Parent
 
-                    # Add the directory to the list if it's not already included
-                    if ($parentDir -notin $processedDirectories) {
-                        $processedDirectories += $parentDir
+                        # Add the directory to the list if it's not already included
+                        if ($parentDir -notin $processedDirectories) {
+                            $processedDirectories += $parentDir
+                        }
                     }
                 }
             }